added more failure checking and cleanup
one helper function to download them all.
also generalized the extraction process.
there is an untar-wrapper that makes it so we can use the same function call for extraction.
delete './steamcmd/' if we couldn't actually create it.
give us more status about the process of downloading and extracting steamcmd (on both platforms)